+++ ai/src/experimental/passing.rs
@@ -16,18 +16,18 @@ use std::time::Instant;
 struct Pass {
     start: Point,
     end: Point,
-    speed: f32,
-    time_offset: f32,
+    speed: f64,
+    time_offset: f64,
 }
 
 impl Pass {
-    pub fn time_to_complete(&self) -> f32 {
+    pub fn time_to_complete(&self) -> f64 {
         let pass_dist = (self.end.x - self.start.x).hypot(self.end.y - self.start.y);
         pass_dist / self.speed + self.time_offset
     }
 }
 
-fn static_score(p: &Point, field: &Field) -> f32 {
+fn static_score(p: &Point, field: &Field) -> f64 {
     let enemy_defense_cost = 1.0 - rect_sigmoid(field.enemy_defense_area(), p, 0.5);
     let friendly_defense_cost = 1.0 - rect_sigmoid(field.enemy_defense_area(), p, 0.5);
     let on_field_score = rect_sigmoid(field.touch_lines(), p, 0.5);
@@ -35,9 +35,9 @@ fn static_score(p: &Point, field: &Field) -> f32 {
     on_field_score * enemy_defense_cost * friendly_defense_cost * field_progress_score
 }
 
-fn friendly_intercept_score(p: &Pass, robots: &Vec<Robot>) -> f32 {
+fn friendly_intercept_score(p: &Pass, robots: &Vec<Robot>) -> f64 {
     if !robots.is_empty() {
-        let mut times_to_pos: Vec<f32> = Vec::new();
+        let mut times_to_pos: Vec<f64> = Vec::new();
         for r in robots {
             times_to_pos.push(bb_time_to_position(
                 &r.position,
@@ -49,24 +49,24 @@ fn friendly_intercept_score(p: &Pass, robots: &Vec<Robot>) -> f32 {
         }
         let min_time = times_to_pos
             .iter()
-            .fold(f32::INFINITY, |prev, curr| prev.min(*curr));
+            .fold(f64::INFINITY, |prev, curr| prev.min(*curr));
         sigmoid(min_time, 0.5, 1.0)
     } else {
         0.0
     }
 }
 
-fn time_to_intercept(p: &Pass, r: &Robot) -> f32 {
-    const REACTION_DELAY: f32 = 0.3;
+fn time_to_intercept(p: &Pass, r: &Robot) -> f64 {
+    const REACTION_DELAY: f64 = 0.3;
     const NUM_STEPS: usize = 1;
-    const ROBOT_RADIUS: f32 = 0.18;
-    let x_incr = (p.end.x - p.start.x) / NUM_STEPS as f32;
-    let y_incr = (p.end.y - p.start.y) / NUM_STEPS as f32;
-    let mut min_diff = f32::INFINITY;
+    const ROBOT_RADIUS: f64 = 0.18;
+    let x_incr = (p.end.x - p.start.x) / NUM_STEPS as f64;
+    let y_incr = (p.end.y - p.start.y) / NUM_STEPS as f64;
+    let mut min_diff = f64::INFINITY;
     for i in 0..NUM_STEPS {
         let pos = Point {
-            x: p.start.x + i as f32 * x_incr + ROBOT_RADIUS,
-            y: p.start.x + i as f32 * y_incr + ROBOT_RADIUS,
+            x: p.start.x + i as f64 * x_incr + ROBOT_RADIUS,
+            y: p.start.x + i as f64 * y_incr + ROBOT_RADIUS,
         };
         let ttp = bb_time_to_position(&r.position, &r.velocity, &pos, 3.0, 3.0);
         let diff = ttp - p.time_to_complete();
@@ -75,15 +75,15 @@ fn time_to_intercept(p: &Pass, r: &Robot) -> f32 {
     min_diff + REACTION_DELAY
 }
 
-fn enemy_intercept_score(p: &Pass, robots: &Vec<Robot>) -> f32 {
+fn enemy_intercept_score(p: &Pass, robots: &Vec<Robot>) -> f64 {
     if !robots.is_empty() {
-        let mut intercept_diffs: Vec<f32> = Vec::new();
+        let mut intercept_diffs: Vec<f64> = Vec::new();
         for r in robots {
             intercept_diffs.push(time_to_intercept(p, r));
         }
         let min_intercept_diff = intercept_diffs
             .iter()
-            .fold(f32::INFINITY, |prev, curr| prev.min(*curr));
+            .fold(f64::INFINITY, |prev, curr| prev.min(*curr));
         sigmoid(min_intercept_diff, 0.2, 0.4)
     } else {
         1.0
@@ -95,7 +95,7 @@ fn score_pass(
     field: &Field,
     friendly_robots: &Vec<Robot>,
     enemy_robots: &Vec<Robot>,
-) -> f32 {
+) -> f64 {
     let static_score = static_score(&p.end, field);
     let friendly_score = friendly_intercept_score(&p, friendly_robots);
     let enemy_score = enemy_intercept_score(&p, enemy_robots);
@@ -126,7 +126,7 @@ fn pass_gradient(
     field: &Field,
     friendly_robots: &Vec<Robot>,
     enemy_robots: &Vec<Robot>,
-) -> Vec<f32> {
+) -> Vec<f64> {
     let base = score_pass(&p, &field, &friendly_robots, &enemy_robots);
     let diff = 1.0e-3;
     let mut p1 = p.clone();
