+++ ai/src/gameplay/tactic.rs
@@ -10,7 +10,7 @@ pub enum Tactic {
 }
 
 impl Tactic {
-    pub fn robot_assignment_cost(&self, robot: &Robot) -> f32 {
+    pub fn robot_assignment_cost(&self, robot: &Robot) -> f64 {
         match self {
             Self::Stop => 0.5,
             Self::Move((p, a)) => (p - &robot.state.position).length(),
