+++ ai/src/geom/point.rs
@@ -3,8 +3,8 @@ use std::ops::{Add, Sub};
 
 #[derive(Clone, Copy, Debug)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 impl Point {
