+++ ai/src/math.rs
@@ -1,13 +1,13 @@
 use super::geom::{Point, Rectangle};
 
 // TODO: Possibly useful later: https://docs.rs/fast-math/latest/fast_math/fn.exp.html
-pub fn sigmoid(x: f32, offset: f32, width: f32) -> f32 {
+pub fn sigmoid(x: f64, offset: f64, width: f64) -> f64 {
     let sig_change_factor = 8.0 / width;
 
     1.0 / (1.0 + (sig_change_factor * (offset - x)).exp())
 }
 
-pub fn rect_sigmoid(rect: Rectangle, p: &Point, width: f32) -> f32 {
+pub fn rect_sigmoid(rect: Rectangle, p: &Point, width: f64) -> f64 {
     let x_offset = rect.centre().x;
     let y_offset = rect.centre().y;
     let x_size = rect.len_x() / 2.0;
